remover arquivo de utilizades e migrar utils para dto
